+++ bika/lims/content/sample.py
@@ -19,7 +19,7 @@ from bika.lims.browser.widgets.datetimewidget import DateTimeWidget
 from bika.lims.browser.widgets import RejectionWidget
 from bika.lims.config import PROJECTNAME
 from bika.lims.content.bikaschema import BikaSchema
-from bika.lims.interfaces import ISample, ISamplePrepWorkflow
+from bika.lims.interfaces import ISample
 from bika.lims.permissions import SampleSample
 from bika.lims.permissions import ScheduleSampling
 from bika.lims.workflow import doActionFor
@@ -331,31 +331,6 @@ schema = BikaSchema.copy() + Schema((
             render_own_label=True,
         ),
     ),
-    StringField('PreparationWorkflow',
-        mode="rw",
-        read_permission=permissions.View,
-        write_permission=permissions.ModifyPortalContent,
-        vocabulary='getPreparationWorkflows',
-        acquire=True,
-        widget=SelectionWidget(
-            format="select",
-            label=_("Preparation Workflow"),
-            visible={'edit': 'visible',
-                     'view': 'visible',
-                     'header_table': 'visible',
-                     'sample_registered': {'view': 'visible', 'edit': 'visible'},
-                     'to_be_sampled':     {'view': 'visible', 'edit': 'invisible'},
-                     'sampled':           {'view': 'visible', 'edit': 'invisible'},
-                     'to_be_preserved':   {'view': 'visible', 'edit': 'invisible'},
-                     'sample_due':        {'view': 'visible', 'edit': 'invisible'},
-                     'sample_received':   {'view': 'visible', 'edit': 'invisible'},
-                     'expired':           {'view': 'visible', 'edit': 'invisible'},
-                     'disposed':          {'view': 'visible', 'edit': 'invisible'},
-                     'rejected':          {'view': 'visible', 'edit': 'invisible'},
-                     },
-            render_own_label=True,
-        ),
-    ),
     ReferenceField('SamplingDeviation',
         vocabulary_display_path_bound = sys.maxsize,
         allowed_types = ('SamplingDeviation',),
@@ -651,7 +626,7 @@ schema['title'].required = False
 
 
 class Sample(BaseFolder, HistoryAwareMixin):
-    implements(ISample, ISamplePrepWorkflow)
+    implements(ISample)
     security = ClassSecurityInfo()
     displayContentsTab = False
     schema = schema
@@ -838,19 +813,6 @@ class Sample(BaseFolder, HistoryAwareMixin):
         workflow = getToolByName(self, 'portal_workflow')
         return workflow.getInfoFor(self, 'review_state')
 
-    def getPreparationWorkflows(self):
-        """Return a list of sample preparation workflows.  These are identified
-        by scanning all workflow IDs for those beginning with "sampleprep".
-        """
-        wf = self.portal_workflow
-        ids = wf.getWorkflowIds()
-        sampleprep_ids = [wid for wid in ids if wid.startswith('sampleprep')]
-        prep_workflows = [['', ''],]
-        for workflow_id in sampleprep_ids:
-            workflow = wf.getWorkflowById(workflow_id)
-            prep_workflows.append([workflow_id, workflow.title])
-        return DisplayList(prep_workflows)
-
     def getSamplePartitions(self):
         """Returns the Sample Partitions associated to this Sample
         """
@@ -928,14 +890,6 @@ class Sample(BaseFolder, HistoryAwareMixin):
     @security.public
     def guard_receive_transition(self):
         return guards.receive(self)
-
-    @security.public
-    def guard_sample_prep_transition(self):
-        return guards.sample_prep(self)
-
-    @security.public
-    def guard_sample_prep_complete_transition(self):
-        return guards.sample_prep_complete(self)
 
     @security.public
     def guard_schedule_sampling_transition(self):